refactor: move canvas drawing logic into separate files Fixes #7
